2019 09 IntCodeComputerCR Is now backed by a List<Long> (no real need for BigInteger or map).  Must get rid of recursive call to avoid stack overflow on part2

+++ src/main/kotlin/no/rodland/advent_2019/Util.kt
@@ -2,7 +2,6 @@ package no.rodland.advent_2019
 
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.runBlocking
-import java.math.BigInteger
 
 class Util
 
@@ -46,8 +45,8 @@ fun getValueFromDeferred(deferred: Deferred<Int>): Int {
     return value
 }
 
-fun getValueFromDeferredList(deferred: Deferred<List<BigInteger>>): List<BigInteger> {
-    var value: List<BigInteger> = emptyList()
+fun getValueFromDeferredList(deferred: Deferred<List<Long>>): List<Long> {
+    var value: List<Long> = emptyList()
     runBlocking { value = deferred.await() }
     return value
 }
